Login

+++ src/containers/auth/LoginForm.tsx
@@ -34,12 +34,14 @@ const LoginForm = () => {
 
       // API 호출
       login(inputEmail as string, inputPw as string)
-        .then((token) => {
+        .then((res) => {
+          let token = localStorage.getItem("tm-token");
+
           if (checkSaveId)
             localStorage.setItem("tm-saved-id", inputEmail as string);
-          // if (checkKeepLogin) localStorage.setItem("tm-token", token as string);
-          localStorage.setItem("tm-token", token as string);
-          if (!token) {
+          if (checkKeepLogin) localStorage.setItem("tm-token", token as string);
+
+          if (token === "undefined") {
             setError("아이디나 비밀번호가 일치하지 않습니다.");
             return;
           } else {
